Rollback history on failures
When calling a session's chat() function, it now uses a commit() and rollback() pattern off the history object.
The history object stores the "finalized_index", which is the last position that has been committed. On a new commit, the index is moved to the end of the history list. On a rollback, the list is truncated.

This is an easy way of doing rollbacks, but it does have some impacts.
- The history object no longer can truncate at any time. It should now only truncate after a commit() or set_history() call. Truncating sooner would make rollback impossible.
- The history object's self.history now contains the non-truncated history. Nothing was accessing this directly, but to be extra clear it is now called `self._history` so that it is unambigiously private.  All access to the history should call the history object's get_messages() function, which returns the truncated history.