Support Hunyuan Video text to video (#3)
* Start adding Hunyuan Video T2V support

It works, but still needs the CLI flags added and documentation.

* comfy: Add CLI flags for Hunyuan video

* comfy: Add denoise settings for hunyuan video

* settings: Add denoise for hunyuan video

* Minor pep8 adjustment

* comfy: Add LoRA support for Hunyuan Video

* comfy: Remove repeated parameters

This makes all parameter mandatory, but it's an internal function with a magic caller, so that makes sense. The reptition was unnecessary and causes risks of falling out of sync.

* comfy: Fix bug where config changes didn't update the defaults

ComfyCaller() was incorrectly listening for mode changes instead of
config changes for rebuilding the defaults. This worked fine when
switching the mode, but changing individual parameters did not
propagate.

* Group all --set/-s overrides with a single config.update event

* README.md: Add documentation for Hunyuan Video

* README.md: Fix headers and update TOC

* README.md: Small cleanups

* README.md: Link to Tencent\'s repo

* README.md: Fix typo