comfy_caller: Remove redundant arguments
The functions to get the defaults were using config.get() with a
default value, but recent changes made it so that get() fails if the
value doesn't exist.  This means the default could never get used, and
so the defautls were redundant and removed.

Some workflow functions also specified defaults, but since an earlier
change made it so that they are never called directly, and the new
caller automatically populates the configs from the defaults, the
value would never be used.

This change removes all the redundant and unused values to simplify
the code.