Event and session improvement (WIP)
The way ChatSessions are handled works now because there is only one
provider, but in the future that will break. This commit is a step
towards fixing that. Along the way, other things were noticed, such
deregistration of events not taking place, which is a problem if
classes are recreated.

- Event subscriptions now automatically unsubscribe on cleanup
- The session title is cleared on `/clear`
- The `mode` is now stored in the v0.2 serialization format
- The session report displays the mode
- ChatHistory() can now be copied and deep copied
- The util command has basic session support (incomplete)