Increase ruff strictness / code cov to 65%
This commit is one of many that aims to make this code base better for
agents.

It turned out a sneaky agent when requested to increase code coverage
omitted a bunch of files from coverage tracking, so we are dropping
temporarily to a 65% requirement.  (We are 67.78%.)  This can then be
built back up.

Not all classes of ruff checks are enabled just yet.  We'll
incrementally start having agents resolve the issues.